Using heuristics in cypher for all nodes and label cardinalities

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/util/statistics/RollingAverage.java
@@ -37,7 +37,7 @@ public class RollingAverage implements Serializable
         this.windowSize = windowSize;
     }
 
-    public void record( int value )
+    public void record( long value )
     {
         averages[0] = (averages[0] * samplesInCurrentWindow + value) / (samplesInCurrentWindow + 1);
